+++ src/main/java/com/vanym/paniclecraft/core/IProxy.java
@@ -1,5 +1,7 @@
 package com.vanym.paniclecraft.core;
 
+import com.vanym.paniclecraft.command.CommandMod3;
+
 public interface IProxy {
     public void preInit(ModConfig config);
     
@@ -8,4 +10,6 @@ public interface IProxy {
     public void postInit(ModConfig config);
     
     public void configChanged(ModConfig config);
+    
+    public CommandMod3 getCommand();
 }